# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/api/test-ci":{"get":{"operationId":"AppController_testCi","parameters":[],"responses":{"200":{"description":""}}}},"/api/users":{"get":{"operationId":"UserController_getUsers","parameters":[],"responses":{"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]}},"/api/users/set-admin-access":{"post":{"operationId":"UserController_setAdminAccess","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetAdminAccessDto"}}}},"responses":{"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]}},"/api/organizations":{"get":{"operationId":"OrganizationController_getOrganizations","summary":"Get all organizations","parameters":[],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Organization"}}}}},"401":{"description":"Unauthorized"}},"tags":["Organization"],"security":[{"bearer":[]}]},"post":{"operationId":"OrganizationController_createOrganization","summary":"Get organization by ID","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrganizationDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Organization"}}}},"401":{"description":"Unauthorized"}},"tags":["Organization"],"security":[{"bearer":[]}]}},"/api/organizations/{id}":{"get":{"operationId":"OrganizationController_getOrganizationById","summary":"Get organization by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Organization"}}}},"401":{"description":"Unauthorized"}},"tags":["Organization"],"security":[{"bearer":[]}]}},"/api/restaurants":{"get":{"operationId":"RestaurantController_getAllRestaurants","parameters":[],"responses":{"401":{"description":"Unauthorized"}},"tags":["Restaurant"],"security":[{"bearer":[]}]}},"/api/restaurants/{organizationId}":{"get":{"operationId":"RestaurantController_getRestaurants","parameters":[{"name":"organizationId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"401":{"description":"Unauthorized"}},"tags":["Restaurant"],"security":[{"bearer":[]}]}},"/api/services/google-maps/resolve-place-photo/{photoReference}":{"get":{"operationId":"GoogleMapsController_resolvePlacePhoto","parameters":[{"name":"photoReference","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Google Maps"]}},"/api/queue/pause/{queueName}":{"post":{"operationId":"QueueController_pause","summary":"Pause a queue","description":"Pause a queue","parameters":[{"name":"queueName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BooleanOperationResult"}}}},"401":{"description":"Unauthorized"}},"tags":["Queue"],"security":[{"JWT":[]}]}},"/api/queue/resume/{queueName}":{"post":{"operationId":"QueueController_resume","summary":"Resume a queue","description":"Resumes a queue","parameters":[{"name":"queueName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BooleanOperationResult"}}}},"401":{"description":"Unauthorized"}},"tags":["Queue"],"security":[{"JWT":[]}]}},"/api/queue/clear/{queueName}":{"post":{"operationId":"QueueController_clear","summary":"Clear queue","description":"Remove all jobs and data from the queue","parameters":[{"name":"queueName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BooleanOperationResult"}}}},"401":{"description":"Unauthorized"}},"tags":["Queue"],"security":[{"JWT":[]}]}},"/api/queue/status/{queueName}":{"get":{"operationId":"QueueController_getStatus","summary":"Get queue status","description":"Returns queue status and information about the jobs","parameters":[{"name":"queueName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Queue status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueueStatusDto"}}}},"401":{"description":"Unauthorized"}},"tags":["Queue"],"security":[{"JWT":[]}]}},"/api/queue-jobs":{"get":{"operationId":"QueueJobController_findAllJobs","parameters":[],"responses":{"200":{"description":""}}}},"/api/auth/login":{"post":{"operationId":"AuthController_login","summary":"Login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailLoginDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessTokenResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["Auth"]}},"/api/auth/admin/login":{"post":{"operationId":"AuthController_loginAdmin","summary":"Login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailLoginDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessTokenResponse"}}}},"401":{"description":"Unauthorized"}},"tags":["Auth"]}},"/api/auth/invitation/create-invitation-link":{"post":{"operationId":"AuthController_createInvitationLink","summary":"Create shareable invitation link","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInvitationLinkDto"}}}},"responses":{"200":{"description":"Success"},"401":{"description":"Unauthorized"}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/auth/invitation/send-email-invitation":{"post":{"operationId":"AuthController_sendEmailInvitation","summary":"Send email invitation to specific people","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InviteUsersByEmailDto"}}}},"responses":{"200":{"description":"Success"},"401":{"description":"Unauthorized"}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/auth/invitation/validate-invitation":{"post":{"operationId":"AuthController_validateInvitation","summary":"Validate invitation & get invitation data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidateInvitationDto"}}}},"responses":{"200":{"description":"Success"},"401":{"description":"Unauthorized"}},"tags":["Auth"]}},"/api/auth/invitation/accept-invitation":{"post":{"operationId":"AuthController_acceptInvitation","summary":"Accept invitation","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AcceptInvitationDto"}}}},"responses":{"200":{"description":"Success"},"401":{"description":"Unauthorized"}},"tags":["Auth"]}},"/api/filters/organizations":{"get":{"operationId":"FiltersController_getOrganizations","parameters":[],"responses":{"200":{"description":""}}}},"/api/invitations":{"get":{"operationId":"InvitationsController_findAll","parameters":[],"responses":{"401":{"description":"Unauthorized"}},"security":[{"bearer":[]}]}}},"info":{"title":"MIDI FRIENDLY API","description":"Midi friendly API documentation","version":"0.0.1","contact":{}},"tags":[{"name":"MIDI-FRIENDLY","description":""}],"servers":[],"components":{"securitySchemes":{"JWT":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SetAdminAccessDto":{"type":"object","properties":{"userId":{"type":"string"},"allowAccess":{"type":"boolean"}},"required":["userId","allowAccess"]},"Address":{"type":"object","properties":{"street":{"type":"string"},"city":{"type":"string"},"country":{"type":"string"},"zip":{"type":"string"}},"required":["street","city","country","zip"]},"Coordinates":{"type":"object","properties":{"latitude":{"type":"number"},"longitude":{"type":"number"}},"required":["latitude","longitude"]},"Organization":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"address":{"$ref":"#/components/schemas/Address"},"coordinates":{"$ref":"#/components/schemas/Coordinates"}},"required":["_id","name","address","coordinates"]},"CreateOrganizationDto":{"type":"object","properties":{}},"BooleanOperationResult":{"type":"object","properties":{"success":{"type":"boolean","example":"true"}},"required":["success"]},"QueueStatusDto":{"type":"object","properties":{"isPaused":{"type":"boolean","example":"false","description":"Is the queue paused?"},"waiting":{"type":"number","example":"0","description":"Amount of waiting jobs in the queue"},"active":{"type":"number","example":"2","description":"Amount of active jobs in the queue"},"completed":{"type":"number","example":"97","description":"Amount of completed jobs in the queue"},"failed":{"type":"number","example":"2","description":"Amount of failed jobs in the queue"},"delayed":{"type":"number","example":"14","description":"Amount of delayed jobs in the queue"},"paused":{"type":"number","example":"0","description":"Amount of paused jobs in the queue"}},"required":["isPaused","waiting","active","completed","failed","delayed","paused"]},"EmailLoginDto":{"type":"object","properties":{"email":{"type":"string","example":"test@example.com"},"password":{"type":"string","example":"password"}},"required":["email","password"]},"AvailableOrganization":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"}},"required":["_id","name"]},"ActiveAccount":{"type":"object","properties":{"_id":{"type":"string","example":"61a4ddc0c3150de171779ee3","description":"Account Id"},"firstName":{"type":"string","description":"First name"},"lastName":{"type":"string","description":"Last name"},"email":{"type":"string","description":"email"},"avatar":{"type":"string","description":"avatar"},"admin":{"type":"boolean","description":"admin access"}},"required":["_id","firstName","lastName","email","avatar","admin"]},"AccessTokenResponse":{"type":"object","properties":{"accessToken":{"type":"string","description":"access token","example":"eyJhbGciOiJIUzI1NiJ9.eyJfaWQiOiI2MWFlNGNmMjkzN2EzNDY4MTNkYmRjNjMiLCJsYXN0TmFtZSI6IkFkbWluIiwiZmlyc3ROYW1lI"},"organizations":{"description":"User organizations","type":"array","items":{"$ref":"#/components/schemas/AvailableOrganization"}},"account":{"description":"Account data","allOf":[{"$ref":"#/components/schemas/ActiveAccount"}]}},"required":["accessToken","organizations","account"]},"CreateInvitationLinkDto":{"type":"object","properties":{"organizationId":{"type":"string"},"expireAt":{"type":"string"},"maxUsage":{"type":"number"},"targetApp":{"type":"string"}},"required":["expireAt","maxUsage","targetApp"]},"InviteUsersByEmailDto":{"type":"object","properties":{"organizationId":{"type":"string"},"expireAt":{"type":"string"},"targetApp":{"type":"string"},"emails":{"type":"array","items":{"type":"string"}}},"required":["expireAt","targetApp","emails"]},"ValidateInvitationDto":{"type":"object","properties":{"invitationId":{"type":"string"},"emailHash":{"type":"string"}},"required":["invitationId"]},"LinkAccountPayload":{"type":"object","properties":{"type":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"userId":{"type":"string"}},"required":["type","email"]},"RegisterAccountPayload":{"type":"object","properties":{"type":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"userId":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"avatar":{"type":"string"}},"required":["type","email","firstName","lastName"]},"AccountPayload":{"type":"object","properties":{"mode":{"type":"string"},"linkPayload":{"$ref":"#/components/schemas/LinkAccountPayload"},"registerPayload":{"$ref":"#/components/schemas/RegisterAccountPayload"}},"required":["mode"]},"AcceptInvitationDto":{"type":"object","properties":{"invitationId":{"type":"string"},"emailHash":{"type":"string"},"account":{"$ref":"#/components/schemas/AccountPayload"}},"required":["invitationId","account"]}}}}